etl_service/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
lxml==4.9.3
html5lib==1.1
